move tests to the same directory; update coverage report
