+++ packages/create-flex-plugin/src/lib/cli.test.ts
@@ -1,5 +1,5 @@
-import cli from '../../src/lib/cli';
-import createFlexPlugin from '../../src/lib/create-flex-plugin';
+import cli from './cli';
+import createFlexPlugin from './create-flex-plugin';
 
 jest.mock('../../src/lib/create-flex-plugin');
