+++ packages/create-flex-plugin/src/utils/github.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
-import * as fs from 'fs';
-import * as path from 'path';
-import * as mkdirp from 'mkdirp';
+import fs from 'fs';
+import path from 'path';
+import mkdirp from 'mkdirp';
 
 export interface GitHubInfo {
     owner: string;
