+++ packages/create-flex-plugin/src/utils/validators.ts
@@ -1,4 +1,4 @@
-import * as inquirer from 'inquirer';
+import inquirer from 'inquirer';
 import { Question } from 'inquirer';
 import { FlexPluginArguments } from '../lib/create-flex-plugin';
 import { error } from './logging';
