+++ packages/create-flex-plugin/tsconfig.json
@@ -5,5 +5,11 @@
     "baseUrl": ".",
     "rootDir": "src"
   },
-  "exclude": ["node_modules", "__tests__", "dist", "templates"]
+  "exclude": [
+    "node_modules",
+    "**/*.test.ts",
+    "**/*.spec.ts",
+    "dist",
+    "templates"
+  ]
 }
