+++ packages/flex-plugin/src/lib/flex-plugin.test.ts
@@ -1,6 +1,5 @@
 import { Flex } from '@twilio/flex-ui/src/FlexGlobal';
-import { FlexPlugin, loadPlugin } from '../../../src';
-import { FlexGlobal } from '../../../src/lib/flex-plugin';
+import { FlexGlobal, FlexPlugin, loadPlugin } from './flex-plugin';
 
 declare var console: Partial<Console>;
