+++ packages/flex-plugin/src/lib/utils/runtime.test.ts
@@ -1,4 +1,4 @@
-import { getAssetsUrl } from '../../../../src';
+import { getAssetsUrl } from './runtime';
 
 describe('runtime', () => {
     describe('getAssetsUrl', () => {
