+++ packages/flex-plugin/tsconfig.json
@@ -5,5 +5,10 @@
     "baseUrl": ".",
     "rootDir": "src"
   },
-  "exclude": ["node_modules", "__tests__", "dist"]
+  "exclude": [
+    "node_modules",
+    "**/*.test.ts",
+    "**/*.spec.ts",
+    "dist"
+  ]
 }
